refactor(core): break catalogs upward edges and mod_loader cycle
Move Adventure into catalogs, background equipment into inventory,
racial HP sources into mechanics, and mod gating API into CatalogSession
so catalogs/platform no longer depend on upper packages.
