+++ tests/core/engine/test_game_engine.py
@@ -1,6 +1,7 @@
 """Тесты game engine."""
 
-from core.character.models import Adventure, Character
+from core.catalogs.adventure import Adventure
+from core.character.models import Character
 from core.engine.game_engine import GameEngine, GameSession
 from core.types import CharacterClass
