+++ tests/core/inventory/test_background_equipment.py
@@ -4,9 +4,9 @@ from pathlib import Path
 
 import pytest
 
-from core.catalogs.backgrounds import get_background_equipment_items
 from core.character.build import build_new_character
 from core.character.storage import persist_character
+from core.inventory.background_equipment import get_background_equipment_items
 from core.types import CharacterBuildParams, CharacterClass
 from tests.creation_helpers import flat_stats
