+++ tests/core/mechanics/test_stats.py
@@ -4,7 +4,8 @@ import random
 
 import pytest
 
-from core.character.models import Adventure, Character
+from core.catalogs.adventure import Adventure
+from core.character.models import Character
 from core.constants import (
     ability_modifier,
     difficulty_class,
